feat: 채팅방에서 강제 나가기 시 소켓 disconnect 및 나가기 api 호출
